refactor(dashboard): address PR review on JobCard and JobLogsSection
- Move formatDuration to shared utils.ts
- Extract JobCardHeader component from JobCard for the header section
- Use formatLogTime helper in JobLogsSection instead of inline toLocaleTimeString
